+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Note that in line with [Django REST Framework policy](http://www.django-rest-framework.org/topics/release-notes/),
 any parts of the framework not mentioned in the documentation should generally be considered private API, and may be subject to change.
 
-## [3.2.0] - pending
+## [Unreleased]
 
 ### Added
 
-* Added support for serializiing complex structures as attributes. For details please reffer to #769
+* Added support for serializing nested serializers as attribute json value introducing setting `JSON_API_SERIALIZE_NESTED_SERIALIZERS_AS_ATTRIBUTE`
+
+### Fixed
+
 * Avoid `AttributeError` for PUT and PATCH methods when using `APIView`
 
 ### Changed
@@ -23,6 +26,7 @@ any parts of the framework not mentioned in the documentation should generally b
 ### Deprecated
 
 * Deprecate `source` argument of `SerializerMethodResourceRelatedField`, use `method_name` instead
+* Rendering nested serializers as relationships is deprecated. Use `ResourceRelatedField` instead
 
 
 ## [3.1.0] - 2020-02-08